placeholders for impls

+++ unitracker-postgres/src/models/group.rs
@@ -2,7 +2,7 @@ use unitracker_chsu::model::groups::Group as ApiGroup;
 pub struct DbGroup {
     pub id: i64,
     pub title: Box<str>,
-    pub course: i8,
+    pub course: i16,
     pub faculty_id: Option<i64>,
     pub chair_id: Option<i64>,
 }
@@ -12,7 +12,7 @@ impl From<ApiGroup> for DbGroup {
         Self {
             id: value.id,
             title: value.title.into_boxed_str(),
-            course: value.course,
+            course: value.course as i16,
             faculty_id: Some(value.faculty.id),
             chair_id: Some(value.chair.id),
         }
